Day 133: Multiplication Table
This is another simple and basic program that would display the multiplication table of variable num (from 1 to 10). I work a full-time job now so I don't have as much time as I did. However, I will stick to my goal and keep on working on this project until I reach 200 days.